:sparkles: New feature added so that preserve data in local storage
